feat: port database from Postgres to SQLite (#14)
* fix: peel double-gzip from TMDB CDN responses

TMDB's CDN intermittently gzip-encodes responses twice (gzip wrapping
gzip wrapping JSON) while advertising only a single Content-Encoding:
gzip. Go's HTTP transport transparently strips one layer, leaving the
inner gzip stream for json.Decode — which fails with
"invalid character '\x1f'" (gzip magic). Every Discover endpoint then
returned 502 and the Discover page showed no movies. Smaller responses
(/movie/{id}) are only single-gzipped, so library and detail pages were
unaffected.

The TMDB client now reads the body and peels any gzip layers Go's
transport left behind. peelGzip is a no-op on plain JSON, so it
self-heals if TMDB stops double-gzipping.


* feat: port database from Postgres to SQLite

- internal/db/db.go on modernc.org/sqlite, WAL pragmas, MaxOpenConns(1)
- goose dialect "sqlite3"; squash 5 migrations into 00001_initial.sql
- sqlc engine "sqlite"; queries moved postgres/ -> sqlite/; $N -> ?;
  ::text casts removed; strftime replacements for to_char/INTERVAL
- row_id SERIAL becomes INTEGER PRIMARY KEY AUTOINCREMENT; id flips to
  TEXT UNIQUE so v3 Radarr-compat row_id integer IDs round-trip and
  never-reuse semantics match Postgres SERIAL (VACUUM-stable)
- internal/api/v3/adapter.go: $1 -> ? in raw row_id<->UUID lookups
- backup.go ported from pg_dump/psql to VACUUM INTO + magic-byte
  validation; restore replaces the .db file with restart-required UX
- internal/core/dbutil: helpers operate on *string/*int64/*time.Time
  directly; IsUniqueViolation against *modernc/sqlite.Error
- DatabaseConfig collapses to {Path}; remove DefaultDBDriver and dead
  sqlite branch in load.go
- 3 Dockerfiles updated: PRISM_DATABASE_PATH=/config/prism.db, drop
  pgx-era ENV and password-file mounts
- int32->int64 / sql.NullString->*string / TIMESTAMPTZ->TEXT ripple
  across ~25 sqlc call sites in 14 packages


---------

+++ internal/core/customformat/service.go
@@ -181,7 +181,7 @@ func (s *Service) SetScores(ctx context.Context, profileID string, scores map[st
 		if err := s.q.SetCustomFormatScore(ctx, dbgen.SetCustomFormatScoreParams{
 			QualityProfileID: profileID,
 			CustomFormatID:   cfID,
-			Score:            int32(score),
+			Score:            int64(score),
 		}); err != nil {
 			return fmt.Errorf("setting score for format %q: %w", cfID, err)
 		}
